Add mock datas for delete component testing
