+++ src/components/pages/DeleteHouse.js
@@ -21,14 +21,14 @@ const DeleteHouse = () => {
     <div className="delete-house">
       <div className="delete-nav">
         <Link to="/">
-          <FontAwesomeIcon icon={faArrowLeft} className="fa-arrow-left" />
+          <FontAwesomeIcon icon={faArrowLeft} className="fa-arrow-left" data-testid="fa-arrow-left" />
         </Link>
         <h1>Delete a house</h1>
       </div>
 
-      <ul className="delete-house-container">
+      <ul className="delete-house-container" data-testid="delete-house-container">
         {houses.map((house) => (
-          <li key={house.id} className="delete-house-items">
+          <li key={house.id} className="delete-house-items" data-testid="delete-house-item">
             <img src={house.image_url} alt={house.name} />
             <div className="delete-items-details">
               Name:
@@ -43,7 +43,7 @@ const DeleteHouse = () => {
               {' '}
               {house.location}
             </div>
-            <button type="button" onClick={() => handleDeleteHouse(house.id)}>
+            <button type="button" onClick={() => handleDeleteHouse(house.id)} data-testid="delete-button">
               Delete
             </button>
           </li>
